一部クラスのリネーム、MapMakerInputからドメインを一部切り出し
