+++ Assets/Projects/CarrySystem/Map/Scripts/StageMapSaver.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: dc3134ccd2884c5da6624a8792f0222a
+timeCreated: 1700382390
